+++ src/Ethereum.php
@@ -34,6 +34,7 @@ class Ethereum implements CryptoBalanceChecker
      */
     public function toWei()
     {
+        $this->unit = 'Wei';
         return $this->balance;
     }
 
@@ -42,6 +43,7 @@ class Ethereum implements CryptoBalanceChecker
      */
     public function toKwei(): float|int
     {
+        $this->unit = 'Kwei';
         return $this->balance / pow(10, 3);
     }
 
@@ -50,6 +52,7 @@ class Ethereum implements CryptoBalanceChecker
      */
     public function toGwei(): float|int
     {
+        $this->unit = 'Gwei';
         return $this->balance / pow(10, 9);
     }
 
@@ -58,6 +61,7 @@ class Ethereum implements CryptoBalanceChecker
      */
     public function toMwei(): float|int
     {
+        $this->unit = 'Mwei';
         return $this->balance / pow(10, 6);
     }