+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Changelog
 
 All notable changes to `crypto-balance-checker` will be documented in this file.
+## 13-08-2023
+- Added support for Bitcoin and fixed some errors.
 ## 12-08-2023
 - Initial working prototype supporting Ethereum.